pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "metapixel"
version = "2.0.0"
description = "Advanced image analysis tool for extracting and categorizing dominant color palettes."
authors = [
  { name="World of Unreal", email="contact@worldofunreal.com" },
]
readme = "README.md"
requires-python = ">=3.8"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dependencies = [
    "opencv-python",
    "numpy",
    "Pillow",
]

[project.urls]
"Homepage" = "https://github.com/worldofunreal/Metapixel"
"Bug Tracker" = "https://github.com/worldofunreal/Metapixel/issues"

[project.scripts]
metapixel = "metapixel.cli:main"
